// frontend-test/src/app/modules/home/modules/features/modules/customizedmealplan/models/targetCalorieMealsDayResult.ts
  export type MealDay= {
    id: number;
    imageType: string;
    title: string;
    readyInMinutes: number;
    servings: number;
    sourceUrl: string;
    imageUrl:string;
  }
  
  export type NutrientsDay= {
    calories: number;
    protein: number;
    fat: number;
    carbohydrates: number;
  }
  